SublimeREPL and SublimeRope updates
Need to look into ways to not have to commit these but still have most
everything work out perfectly. I've attempted this in the past to
limited success, but it is now a goal to no longer commit Package
Control packages - only those that are part of the Sublime Text 2
install.